# xts: reduce tweak allocations